requirements.txt

django>=4.0
pandas
scikit-learn
matplotlib
gunicorn
whitenoise
joblib
numpy
Pillow
dj-database-url
